requirements.txt
openai==1.107.0
python-decouple==3.8
beautifulsoup4==4.13.5